chore(webapp): refactor QueryInput into its own component with tests
