+++ webapp/javascript/components/QueryInput/QueryInput.module.scss
@@ -1,0 +1,4 @@
+.executeButton {
+  flex-grow: 1;
+  color: var(--ps-immutable-white);
+}
